feat(widget): create list-styled containers with `list::container()`
